+++ src/engine/operator.c
@@ -2,7 +2,7 @@
 
 #include "operator.h"
 
-Operator op_get_function(char *name, unsigned int arity)
+Operator op_get_function(char *name, size_t arity)
 {
     return (Operator){
         .name = name,
